feat(modules): add solve for ems
