+++ src/modules/esm.mjs
@@ -1,0 +1,54 @@
+import path from 'path';
+import { release, version } from 'os';
+import { createServer as createServerHttp } from 'http';
+import { dirname } from 'path';
+import { fileURLToPath } from 'url';
+import { createRequire } from 'node:module';
+import './files/c.js';
+
+const require = createRequire(import.meta.url);
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+
+const random = Math.random();
+
+let unknownObject;
+
+if (random > 0.5) {
+  unknownObject = require('./files/a.json');
+} else {
+  unknownObject = require('./files/b.json');
+}
+
+// // Second option
+// if (random > 0.5) {
+//   await import('./files/a.json', { assert: { type: 'json' } }).then((file) => {
+//     unknownObject = file.default;
+//   });
+// } else {
+//   await import('./files/b.json', { assert: { type: 'json' } }).then((file) => {
+//     unknownObject = file.default;
+//   });
+// }
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+console.log(`Path to current file is ${__filename}`);
+console.log(`Path to current directory is ${__dirname}`);
+
+const myServer = createServerHttp((_, res) => {
+  res.end('Request accepted');
+});
+
+const PORT = 3000;
+
+console.log(unknownObject);
+
+myServer.listen(PORT, () => {
+  console.log(`Server is listening on port ${PORT}`);
+  console.log('To terminate it, use Ctrl+C combination');
+});
+
+export { unknownObject, myServer };
